2010-02-03 15:01 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/contrib/hbct/video.c     * pacified DJGPP warning
  * harbour/external/jpeg/Makefile
    * define HAVE_STDLIB_H - it's necessary for exit() function used in
      jerror.c

+++ harbour/contrib/hbct/video.c
@@ -300,7 +300,7 @@ HB_FUNC( VIDEOTYPE )
 HB_FUNC( SETFONT )
 {
    const char *font = hb_parcx( 1 );
-   int len = hb_parclen( 1 );
+   unsigned len = ( unsigned ) hb_parclen( 1 );
    int area = hb_parni( 2 );
    int offset = 0;
    int count = 256;
